requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.1
ffmpeg-python==0.2.0
future==0.18.2
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.31
py==1.11.0
pyee==9.0.4
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
